+++ compile/index.js
@@ -1,22 +1,22 @@
-const { _myNewPackage } = require('./mnp')
+const { _email } = require('./email')
 
 /**
- * {{ description }}
- * @param {!_myNewPackage.Config} config Options for the program.
+ * Send e-mails by direct connection to recepient's SMPT server.
+ * @param {!_email.Config} config Options for the program.
  * @param {boolean} [config.shouldRun=true] A boolean option. Default `true`.
  * @param {string} [config.text] A text to return.
  * @return {Promise<string>}
  */
-function myNewPackage(config) {
-  return _myNewPackage(config)
+function email(config) {
+  return _email(config)
 }
 
-module.exports = myNewPackage
+module.exports = email
 
 /* typal types/index.xml namespace */
 /**
- * @typedef {_myNewPackage.Config} Config `＠record` Options for the program.
- * @typedef {Object} _myNewPackage.Config `＠record` Options for the program.
+ * @typedef {_email.Config} Config `＠record` Options for the program.
+ * @typedef {Object} _email.Config `＠record` Options for the program.
  * @prop {boolean} [shouldRun=true] A boolean option. Default `true`.
  * @prop {string} [text] A text to return.
  */