+++ test/context/index.js
@@ -1,7 +1,7 @@
 import { join } from 'path'
 import { debuglog } from 'util'
 
-const LOG = debuglog('my-new-package')
+const LOG = debuglog('@artdeco/email')
 
 /**
  * A testing context for the package.
@@ -29,16 +29,4 @@ export default class Context {
   async _destroy() {
     LOG('destroy context')
   }
-  static get BIN() {
-    return BIN
-  }
-}
-
-let BIN = 'src/BIN'
-if (process.env.ALAMODE_ENV == 'test-build') {
-  console.log('Testing build bin...')
-  BIN = 'build/bin/mnp'
-} else if (process.env.ALAMODE_ENV == 'test-compile') {
-  console.log('Testing compile bin...')
-  BIN = 'compile/bin/mnp'
 }