+++ test/mask/bin.js
@@ -1,8 +1,0 @@
-import makeTestSuite from '@zoroaster/mask'
-import Context from '../context'
-
-export default makeTestSuite('test/result/bin', {
-  fork: {
-    module: Context.BIN,
-  },
-})