+++ types/index.js
@@ -1,9 +1,0 @@
-export {}
-
-/* typal types/index.xml namespace */
-/**
- * @typedef {_myNewPackage.Config} Config `＠record` Options for the program.
- * @typedef {Object} _myNewPackage.Config `＠record` Options for the program.
- * @prop {boolean} [shouldRun=true] A boolean option. Default `true`.
- * @prop {string} [text] A text to return.
- */